refactor: move components to correct folders
