+++ src/components/gallery/JoinCommunity.tsx
