+++ src/components/home/CircularText.tsx
